+++ acme/textui/ACMEContainerTree.py
@@ -212,8 +212,8 @@ class ACMEContainerTree(Container):
 		if resource:
 			jsns = commentJson(resource.asDict(sort = True), 
 							explanations = self.app.attributeExplanations,	# type: ignore [attr-defined]
-							getAttributeValueName = CSE.validator.getAttributeValueName,		# type: ignore [attr-defined]
-							width = (self.resourceView.size[0] - 2) if self.resourceView.size[0] > 0 else 9999)		# type: ignore [attr-defined]
+							getAttributeValueName = lambda a, v: CSE.validator.getAttributeValueName(a, v, resource.ty if resource else None),		# type: ignore [attr-defined]
+							width = (self.resourceView.size[0] - 2) if self.resourceView.size[0] > 0 else 9999)										# type: ignore [attr-defined]
 			
 			# Update the requests view
 			self._update_requests(resource.ri)